layout/style/test/chrome/test_moz_document_usability.html

<!doctype html><html><head><meta charset="UTF-8">
<title>@-moz-document usability test (Bug 1035091)</title>
<!-- Per https://bugzilla.mozilla.org/show_bug.cgi?id=1035091
     @-moz-document rules can only be used in user and UA style sheets
     until further notice. -->

<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css">

<style>
  #green  { color: green }
  #author { color: green }
  #user   { color: red }
  #agent  { color: red }
</style>
</head><body>

<p id="green">All text should be green.</p>
<p id="author">All text should be green.</p>
<p id="user">All text should be green.</p>
<p id="agent">All text should be green.</p>

<script>
    SimpleTest.waitForExplicitFinish();
    var wu =
        window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
              .getInterface(Components.interfaces.nsIDOMWindowUtils);

    var ios =
        Components.classes["@mozilla.org/network/io-service;1"]
        .getService(Components.interfaces.nsIIOService);

    function makeSheet(whom, color) {
        return ios.newURI(
          "data:text/css,@-moz-document regexp('.*test_moz_document.*') {" +
          "%23" + whom + " { color: " + color + "!important } }",
          null, null);
    }
    wu.loadSheet(makeSheet("author", "red"),   wu.AUTHOR_SHEET);
    wu.loadSheet(makeSheet("user",   "green"), wu.USER_SHEET);
    wu.loadSheet(makeSheet("agent",  "green"), wu.AGENT_SHEET);

    addLoadEvent(function () {
        var green  = getComputedStyle(document.getElementById("green"),  "").color;
        var author = getComputedStyle(document.getElementById("author"), "").color;
        var user   = getComputedStyle(document.getElementById("user"),   "").color;
        var agent  = getComputedStyle(document.getElementById("agent"),  "").color;

        is(author, green, "@-moz-document should not be honored in author sheets");
        is(user,   green, "@-moz-document should be honored in user sheets");
        is(agent,  green, "@-moz-document should be honored in agent sheets");

        SimpleTest.finish();
    });
</script>
</body></html>
